+++ api/routes/userRouter.js
@@ -1,8 +1,6 @@
 const express = require('express');
 const router = express.Router();
 
-const userRouter = require('./userRouter');
-
-router.use('/user', userRouter);
+const { AuthMiddleware } = require('../utils/AuthMiddleware');
 
 module.exports = router;
